Move Palpable* to separate files
