+++ osu.Game.Rulesets.Catch/Objects/CatchHitObject.cs
@@ -105,14 +105,6 @@ namespace osu.Game.Rulesets.Catch.Objects
         protected override HitWindows CreateHitWindows() => HitWindows.Empty;
     }
 
-    /// <summary>
-    /// Represents a single object that can be caught by the catcher.
-    /// </summary>
-    public abstract class PalpableCatchHitObject : CatchHitObject
-    {
-        public override bool CanBePlated => true;
-    }
-
     public enum FruitVisualRepresentation
     {
         Pear,
